refactor admin image conversion
